fix lỗi nhỏ

+++ view/ql_dichvu/ql_dichvu_handle.py
@@ -2,7 +2,7 @@ import sys
 import os
 project_path = os.path.abspath(os.path.join(os.path.dirname(__file__), "../../")) 
 sys.path.append(project_path)
-print(project_path) #vi tri tu file hien tai toi root !! phai toi root moi dung duoc
+from view.css import *
 from PyQt6 import QtCore, QtGui, QtWidgets
 from PyQt6.QtWidgets import QTableWidgetItem, QMessageBox
 from view.ql_dichvu.ql_dichvu import Ui_Form
@@ -27,10 +27,11 @@ class ql_dichvu_ui(QtWidgets.QWidget, Ui_Form):
         self.dis_pla.verticalHeader().setVisible(False)
         self.dis_pla.setSelectionMode(QtWidgets.QTableView.SelectionMode.SingleSelection)
         self.dis_pla.setSelectionBehavior(QtWidgets.QTableView.SelectionBehavior.SelectRows)
+        css(self)
+
         # kết nối db
         self.conn = sqlite3.connect("db/hotel7-3.db")
         self.cursor = self.conn.cursor()
-
         self.show_all()
 
         # sự kiện
@@ -40,7 +41,7 @@ class ql_dichvu_ui(QtWidgets.QWidget, Ui_Form):
         self.edi_btn.clicked.connect(self.update_item)
         self.del_btn.clicked.connect(self.delete_item)
         self.dis_pla.selectionModel().selectionChanged.connect(self.selec_row)
-    
+
     def show_err(self):
         msg = QMessageBox()
         msg.setIcon(QMessageBox.Icon.Critical)
@@ -55,8 +56,6 @@ class ql_dichvu_ui(QtWidgets.QWidget, Ui_Form):
             self.dis_pla.setItem(row_index, 0, QTableWidgetItem(str(row_data.dv_id)))
             self.dis_pla.setItem(row_index, 1, QTableWidgetItem(str(row_data.ten)))
             self.dis_pla.setItem(row_index, 2, QTableWidgetItem(str(row_data.gia)))
-        print(data)
-        print("Hien thi")
     
     def insert_item(self):
         # giới hạn quyền
@@ -66,13 +65,14 @@ class ql_dichvu_ui(QtWidgets.QWidget, Ui_Form):
         # id = self.in_id.text()
         ten = self.in_ten.text()
         gia = self.in_price.text()
-        
+
+        if ten=="" or gia=="":
+            msg = QMessageBox();msg.setWindowTitle("Cảnh báo");msg.setText("Vui lòng điền đủ thông tin dịch vụ!");msg.setIcon(QMessageBox.Icon.Warning);msg.exec()
+            return
         try:
             self.cursor.execute("insert into dich_vu (ten_dv, gia) values(?, ?)", (ten, gia))
             self.conn.commit()
-            print("OK")
         except:
-            print("no ok")
             msg = QMessageBox()
             msg.setIcon(QMessageBox.Icon.Critical)
             msg.setWindowTitle("Lỗi")
@@ -130,18 +130,18 @@ class ql_dichvu_ui(QtWidgets.QWidget, Ui_Form):
             self.par.gioi_han_quyen()
             return
         id = self.selec_row()
+        if id==None:
+            msg = QMessageBox();msg.setWindowTitle("Cảnh báo");msg.setText("Hãy chọn dịch vụ muốn xóa!");msg.setIcon(QMessageBox.Icon.Warning);msg.exec()
+            return
         self.cursor.execute("delete from dich_vu where dv_id=?", (id,))
         self.conn.commit()
-        print("xoa thanh cong")
         self.show_all()
     def search_item(self):
         id = self.in_sea.text()
         if id=="":
-            print("du lieu khong hop le")
             self.show_err()
             return
         data = self.dv_dao.search_dich_vu(id)
-        print(data)
         self.dis_pla.setRowCount(0)
         for row_index, row_data in enumerate(data):
             self.dis_pla.insertRow(row_index)
@@ -149,14 +149,11 @@ class ql_dichvu_ui(QtWidgets.QWidget, Ui_Form):
             self.dis_pla.setItem(row_index, 1, QTableWidgetItem(str(row_data.ten)))
             self.dis_pla.setItem(row_index,2, QTableWidgetItem(str(row_data.gia)))
 
-        print("Tim kiem thanh cong")
-
     def selec_row(self):
         row = self.dis_pla.currentRow()
         if row<0:
             return
         id = self.dis_pla.item(row, 0).text()
-        print(id)
         return id
 
 if __name__ == "__main__":
